doan-doctruyen

+++ truyen/app/Http/Controllers/Manage/ProfileController.php
@@ -28,38 +28,43 @@ class ProfileController extends Controller
             'image.required' => 'Bắt buộc gắn ảnh',
             'image.mimes' => 'Hình ảnh phải là một loại tệp: png, jpg.',
         ];
-        $validatedData =$request->validate([
+        $rules = [
             'image' => 'mimes:jpeg,jpg,png|required|max:10000',
-        ],$messages);
+        ];
 
-        $file = $request->image;
+        $validator = Validator::make($request->all(), $rules, $messages);
+        if ($validator->fails()) {
+            return redirect()->route('manage.profile.index')->withErrors($validator)->withInput();
+        } else {
+            $file = $request->image;
 
-        $file->move('images/avatar' , $file->getClientOriginalName());
+            $file->move('images/avatar' , $file->getClientOriginalName());
 
-        $data= array(
-            'image' => $file->getClientOriginalName(),
-        );
+            $data= array(
+                'image' => $file->getClientOriginalName(),
+            );
 
-        if(auth()->user()->image == ''){
-
-            User::where('id',auth()->user()->id)->update($data);
-
-            $request->session()->flash('success' , 'Image Avatar has been updated');
-        }else{
-            if(auth()->user()->image == $file->getClientOriginalName()){
+            if(auth()->user()->image == ''){
 
                 User::where('id',auth()->user()->id)->update($data);
 
                 $request->session()->flash('success' , 'Image Avatar has been updated');
             }else{
-                $Image = public_path("/images/avatar/" . auth()->user()->image); //Finding users previous picture
+                if(auth()->user()->image == $file->getClientOriginalName()){
 
-                if(file_exists($Image)){ //If it exits, delete it from folder
-                    unlink($Image);
+                    User::where('id',auth()->user()->id)->update($data);
+
+                    $request->session()->flash('success' , 'Image Avatar has been updated');
+                }else{
+                    $Image = public_path("/images/avatar/" . auth()->user()->image); //Finding users previous picture
+
+                    if(file_exists($Image)){ //If it exits, delete it from folder
+                        unlink($Image);
+                    }
+                    User::where('id',auth()->user()->id)->update($data);
+
+                    $request->session()->flash('success' , 'Image Avatar has been updated');
                 }
-                User::where('id',auth()->user()->id)->update($data);
-
-                $request->session()->flash('success' , 'Image Avatar has been updated');
             }
         }
 
@@ -77,24 +82,26 @@ class ProfileController extends Controller
         $messages = [
             'name.required' => 'Tên bắt buộc nhập',
             'name.max' => 'Từ khóa không được vượt quá 30 ký tự',
-            'image.required' => 'Bắt buộc gắn ảnh',
-            'image.mimes' => 'Hình ảnh phải là một loại tệp: png, jpg.'
         ];
-        $validatedData =$request->validate([
+        $rules = [
             'name' => 'required|max:30|',
-            'image' => 'mimes:jpeg,jpg,png|required|max:10000',
-        ],$messages);
+        ];
 
-        $data = ([
-            'name' => $request->name,
-        ]);
+        $validator = Validator::make($request->all(), $rules, $messages);
+        if ($validator->fails()) {
+            return redirect()->route('manage.profile.index')->withErrors($validator)->withInput();
+        } else {
+            $data = ([
+                'name' => $request->name,
+            ]);
 
-        $id_user = auth()->user()->id;
+            $id_user = auth()->user()->id;
 
-        if(User::where('id',$id_user)->update($data)){
-            $request->session()->flash('success', 'Proflie has been updated');
-        }else{
-            $request->session()->flash('error' , ' Error!!!');
+            if(User::where('id',$id_user)->update($data)){
+                $request->session()->flash('success', 'Proflie has been updated');
+            }else{
+                $request->session()->flash('error' , ' Error!!!');
+            }
         }
 
         return redirect()->route('manage.profile.index');
